web: Move dolt query hooks
