+++ web/hooks/useDoltQueryBuilder/tests/types.ts
@@ -1,0 +1,3 @@
+export type Test<T> = { args: T; expected: string; desc: string };
+
+export type Tests<T> = Array<Test<T>>;
