+++ web/hooks/useDoltQueryBuilder/types.ts
@@ -1,0 +1,4 @@
+export type ReturnType = {
+  generateQuery: () => string;
+  isPostgres: boolean;
+};
